refactor(bpa): split the validating decorator from the spool rig
Dispatcher::spool is now a pure store rig — spool(stream, token) ->
Result<(Arc<str>, usize), ConcatError>: a spawned save_stream task
fed through the bounded channel by a pump over the caller's borrowed
stream, and nothing else. It knows nothing about validation, so any
input door can drive it; the originate doors are the next customers.

The validation moves to the doors as dispatcher::validate — the
TailReceiver machinery renamed to what it really is: a
ValidatingReceiver (with ValidationFailure as its verdict) that an
input door wraps over its arrival stream before handing it to the
spool. It yields the resident head first, feeds the PayloadTail
checks and the deferred payload-BIB digests as segments flow, and
settles the categorised verdict from finish() once the spool
returns. Shared in the dispatcher tree deliberately: the raw
originate door performs the same validation when it streams.

Two contract shifts ride the split. Cancellation is solely the spool
pump's job — the token comes out of the decorator (its internal race
was an artifact of running task-side over the channel; the pump's
CancellableReceiver covers mid-park cancels for every caller), and a
cancelled drain simply stops being pulled, settling Truncated. And
the door now settles the verdict and owes the discard: a save the
post-stream validation rejects is deleted by the door's settle arm,
exactly as a canceller already owed the discard of a save racing the
cancel — the task can no longer judge either case.

Style pass on the touched dispatcher files: the super::* globs are
replaced with explicit three-block imports, and the inline
multi-segment paths they hid (CancellationToken::new,
otel_metrics::* labels, eid::NodeId, channel::bounded) are hoisted
into the use blocks.

- tests: the decorator suite moves to validate.rs intact; the
  token-cancel unit retires with the token — the cancel-unblocks-a-
  parked-transfer property stays pinned end to end by
  early_route_drop_never_awaits_the_payload. Behaviour-neutral:
  lib and pipeline suites pass unchanged.
