+++ bpa/CHANGELOG.md
@@ -18,7 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 - `Bytes` implements `stream::Receiver<Segment>` — a whole in-memory buffer is a one-`Final`-segment stream, zero-copy — and `stream::buffer_stream` assembles a segment stream into a contiguous buffer, enforcing `total_len` exactness and 32-bit addressability via the new `stream::BufferError` (which converts into `cla::Error`/`services::Error` through the new `PayloadUnaddressable` variants).
 
 ### Changed
-- The streamed ingress drain is concurrent, behind the store's streaming write seam. `Store::save_stream(&mut dyn Receiver<Segment>)` is the seam at its target signature (`docs/streaming_pipeline_design.md` §3): one stream carries the whole bundle — `TailReceiver` yields the resident head as its first segment, then validates the tail as it flows — and the interim body spools via `concat_stream` and commits through the backends' whole-buffer `save`; the storage tranche swaps the body and the backend trait with no caller changing. A spawned task owns the drain end-to-end (fed by a bounded-channel pump over the CLA stream) while the gate decisions — Ingress chain and route lookup — run alongside; a gate drop cancels the spool mid-stream, so a rejected bundle never waits for its payload and is never persisted (a save racing the cancel, or failing post-stream validation, is discarded: the streaming contract's stage-then-discard ordering). The size cap is enforced against the *declared* wire size the moment the headers parse — an over-cap bundle refuses before a single payload byte drains, and a producer exceeding its declaration trips the framing checks rather than a byte-count. The recovery paths take the same door: reassembly and restart-orphan re-admission hand their bytes as the bundle stream and the spool saves an admitted bundle fresh — there is no caller-stored special case, and each caller deletes its superseded copy after the pipeline settles.
+- The streamed ingress drain is concurrent, behind the store's streaming write seam. `Store::save_stream(&mut dyn Receiver<Segment>)` is the seam at its target signature (`docs/streaming_pipeline_design.md` §3): one stream carries the whole bundle — a validating decorator (`ValidatingReceiver`) yields the resident head as its first segment, then validates the tail as it flows — and the interim body spools via `concat_stream` and commits through the backends' whole-buffer `save`; the storage tranche swaps the body and the backend trait with no caller changing. `Dispatcher::spool` is a validation-blind store rig (a spawned `save_stream` task fed by a bounded-channel pump over the door's decorated stream) running while the gate decisions — Ingress chain and route lookup — run alongside; a gate drop cancels the spool mid-stream, so a rejected bundle never waits for its payload and is never persisted (a save racing the cancel, or failing post-stream validation, is discarded: the streaming contract's stage-then-discard ordering). The size cap is enforced against the *declared* wire size the moment the headers parse — an over-cap bundle refuses before a single payload byte drains, and a producer exceeding its declaration trips the framing checks rather than a byte-count. The recovery paths take the same door: reassembly and restart-orphan re-admission hand their bytes as the bundle stream and the spool saves an admitted bundle fresh — there is no caller-stored special case, and each caller deletes its superseded copy after the pipeline settles.
 - Routing is decided at the ingress gate and is the decision of record for fresh arrivals: the RIB lookup runs on the resident headers concurrently with the payload drain, a destination resolving to an explicit `Drop` route rejects the bundle without persisting or awaiting its payload (reported like any gated drop when the route carries a reason; silent when it does not), and the commit executes the carried decision directly — fresh arrivals no longer transit `DispatchPending`, which now belongs to the originate and re-dispatch paths (parks, polls, sweeps, restart, transfer outcomes). A decision that proves stale during the drain lands in a failure arm whose park re-checks the gate-time table snapshot and re-enters dispatch, and a crash between the metadata insert and execution recovers through restart's `Dispatching` arm — both machinery that already existed. Fresh-arrival execution is bounded by CLA session concurrency rather than the dispatch pool, in line with the plan for backpressure to live at the CLAs, services, and stores.
 - **BREAKING:** the egress queue model is totalised, and the policy traits take the redesign's names: `EgressPolicy` is `FlowControllerFactory` and `EgressController` is `FlowController` (`EgressQueue` keeps its name — in the general case it is a queue, feeding one lane directive). `FlowControllerFactory::queue_count` returns `NonZeroU32` — every policy has a queue for every bundle to classify into, and queue indices are the policy's own naming (index 0 is only guaranteed to exist; relative priority is a policy-internal decision) — and `FlowControllerFactory::classify` is deleted: the per-peer `FlowController` owns the queue assignment (`queue_for()`, an index in `0..queue_count()`; an out-of-range index is clamped to queue 0 and logged), so the hot forwarding path touches no shared policy state — the first recorded shift of `docs/policy_subsystem_redesign.md`'s FlowController model. `FlowController::forward` and `BundleStatus::ForwardPending` carry the plain index, through the storage codecs (SQLite/PostgreSQL store the queue parameter unconditionally). The `Option` remains where it is meaningful — the lane directive keying `new_controller`'s queue set: `Some(n)` transmits pinned to declared lane `n`, `None` (always present) on the next free lane.
 - **BREAKING:** `storage::MetadataStorage::replace` is removed, and with it the last non-status rewrite of a stored record: a record's non-status metadata is fixed at insert, and thereafter only its status moves (`update_status`/`swap_status`/`tombstone_*`). Backends drop their `replace` implementations.
