+++ bpa/docs/streaming_pipeline_design.md
@@ -4,7 +4,7 @@
 | --- | --- |
 | **Component** | BPA — Bundle data I/O, spool-based streaming |
 | **Scope** | Transformer-based streaming, spool commit model, sequential-only storage |
-| **Status** | Living doc — partially implemented (streaming parser, CLA segment delivery, unified ingress pipeline with pre-drain gate + validating `TailReceiver` drain landed; storage / egress / tee'd spool pending) |
+| **Status** | Living doc — partially implemented (streaming parser, CLA segment delivery, unified ingress pipeline with pre-drain gate + `ValidatingReceiver` validating drain landed; storage / egress / tee'd spool pending) |
 | **Related** | `queue_architecture.md`, `storage_subsystem_design.md`, Editor (`bpv7/src/editor.rs`) |
 
 ## Implementation status (2026-09-03)
@@ -15,9 +15,9 @@ This is a living document: the sections below mix shipped behaviour with design 
 
 - Streaming parser — `bpv7::parse::BundleParser` (`push` / `finish`), `ParserProgress::{NeedMore, Ready, Partial}`, one-shot sugar `bpv7::parse::parse(Bytes) -> Parsed`. `Partial { consumed, tail: PayloadTail }` is the deferred-payload signal: `push` returns it at the payload boundary and the caller drains the remaining payload through `PayloadTail` (CRC + termination + anti-smuggling checks) without re-buffering headers. Non-canonical CBOR is a hard parse error (§5.2.2). Keyless BPSec structural checks are `bpsec::{bib,bcb}::OperationSet::check`.
 - CLA segment delivery — `cla::Segment::{Next, Final}` and the streamed-only `Sink::dispatch(&mut dyn Receiver<Segment>)` (the buffered/streamed trait pairs were collapsed; no `Bytes` variant remains). Stream traits `Sender<T>` / `Receiver<T>` live in `bpa::stream` with adapters over `hardy_async::channel`.
-- Streamed ingress drain — `dispatcher::ingress` drives the parser per `Segment` and drains the bundle through the validating `TailReceiver` (below, yielding the resident head first so one receiver carries the whole bundle) into one `Store::save_stream(&mut dyn Receiver<Segment>)` call: the store owns the spool, so the pipeline already assumes an asynchronous streaming store, and the reject paths already stage-then-discard as the streaming contract requires.
+- Streamed ingress drain — `dispatcher::ingress` drives the parser per `Segment` and drains the bundle through a `ValidatingReceiver` (below, yielding the resident head first so one receiver carries the whole bundle) into one `Store::save_stream(&mut dyn Receiver<Segment>)` call: the store owns the spool, so the pipeline already assumes an asynchronous streaming store, and the reject paths already stage-then-discard as the streaming contract requires.
 - Pre-drain gate — `bundle::parse::parse_headers` + `HeaderVerify::gate_reason`: lifetime/hop/expiry rejection, the config-gated RFC 9171 validity checks, and the Ingress filter chain all run on the accumulation buffer before the payload drain (§5.4's early reject in interim form; link-layer reach-back per §1.3 rides the CLA's streaming dispatch path). A gate drop is pre-custody and never persisted. The RIB lookup also runs at the gate as the routing decision of record — an explicit Drop route rejects here without awaiting the payload, cancelling the concurrent spool mid-stream, and the commit executes the carried decision (§5.7's routing-at-commit, both halves).
-- `TailReceiver` validating drain (§5.5's pull-through) — `bundle::tail::TailReceiver` wraps the CLA stream as a `Receiver<Segment>` decorator: as each segment flows through it feeds the sync `PayloadTail` checks (payload CRC, block/outer breaks, anti-smuggling) plus one incremental digest per deferred payload BIB, then yields the same segment onward; the categorised verdict (`Truncated` / `Invalid` / `IntegrityFailed`) is read from `finish()` once drained. Layering is as designed: `PayloadTail` and the digest framing are `bpv7`'s (no_std, sans-IO), the async wrapper is bpa's.
+- `ValidatingReceiver` validating drain (§5.5's pull-through) — the dispatcher's shared `Receiver<Segment>` decorator, wrapped by an input door over its arrival stream: as each segment flows through it feeds the sync `PayloadTail` checks (payload CRC, block/outer breaks, anti-smuggling) plus one incremental digest per deferred payload BIB, then yields the same segment onward; the categorised verdict (`Truncated` / `Invalid` / `IntegrityFailed`) is read from `finish()` after the spool returns, and the door owns the discard of a save the verdict rejects. Layering is as designed: `PayloadTail` and the digest framing are `bpv7`'s (no_std, sans-IO), the async wrapper is bpa's — and validation is the door's, so `Dispatcher::spool` stays a pure store rig any door can drive (the originate doors decorate their own streams).
 - Deferred payload-BIB verification is streaming — `checks::begin_payload_verification` begins a `bpsec::bib::Verifier` per BIB the header verify deferred, *inside `parse_headers`' keyed scope* (`HeaderVerify::deferred_verifiers`): the `!Send` `KeySource` is resolved once per bundle and never crosses an `await`; only the `Send` verifiers — copied key material, the recorded exception documented on `bib::Verifier` — ride the drain. `checks::verify_payload` and `finalize_with_provider` are deleted, resolving the key-material re-adjudication this block previously flagged.
 - Unified ingress pipeline, one metadata write — the CLA, reassembly, and restart in-feeds share `process_received_bundle`; the single `insert_metadata` at `Dispatching` carries the Ingress chain's classifier deltas (`New` is an in-memory marker that never persists), and the dispatch send's CAS to `DispatchPending` is the queue commit.
 - No editing on input — the bundle is stored exactly as received; §5.1.1 failure-drops and unrecognised-deletable removals are *scheduled* in `BundleMetadata::to_remove` and applied per attempt at the output doors (the egress rewrite head in `forward.rs`, the deliver strip). Captured payload-BCB op-sets are likewise never spent at ingress — delivery re-derives them from the stored bytes (the §6.1.5 transformers make that streaming).
@@ -28,7 +28,7 @@ This is a living document: the sections below mix shipped behaviour with design 
 **Interim (works, not yet the target shape)**
 
 - The spool's destination is still RAM: `Store::save_stream` accumulates its segment stream (bounded by `max_bundle_size`) and commits via `BundleStorage::save(Bytes)` on the final segment. The seam now lives *inside the store wrapper*, behind the target signature — one `Receiver<Segment>` in, storage name out — so the storage-spool tranche swaps `save_stream`'s body (and the backend trait) for the streamed write (§3) without touching pipeline code: every stage upstream already sees a plain `Receiver<Segment>`, and ingress already deletes a staged save whose post-stream validation fails (the discard half of the contract).
-- The drain is concurrent: a spawned task owns the `TailReceiver` and the spool (fed by a bounded-channel pump over the borrowed CLA stream), racing the gate decisions, with a cancel token aborting the spool on a chain or route drop — §5.7's shape, landed. What remains interim is only the spool's RAM destination above.
+- The drain is concurrent: `Dispatcher::spool` — a spawned store-side task fed by a bounded-channel pump over the door's decorated stream — races the gate decisions, with a cancel token aborting the spool on a chain or route drop — §5.7's shape, landed. What remains interim is only the spool's RAM destination above.
 
 **Pending**
