+++ bpa/src/dispatcher/mod.rs
@@ -17,6 +17,7 @@ mod reassemble;
 mod report;
 mod restart;
 mod spool;
+mod validate;
 
 // The default bound on a single reassembled bundle. Streaming producers
 // dissolved the transport-level caps that used to bound ingress implicitly,
